refactor(core): extract custom state feature
- new custom pseudo state feature
- move `custom-states` & `state-validators` to custom-state helper file
