+++ packages/core/src/features/st-custom-state.ts
@@ -1,0 +1,36 @@
+import { createFeature } from './feature';
+import {
+    stateDiagnostics,
+    parsePseudoStates,
+    transformPseudoClassToCustomState,
+    booleanStateDelimiter,
+    stateMiddleDelimiter,
+    stateWithParamDelimiter,
+    createBooleanStateClassName,
+    createStateWithParamClassName,
+    systemValidators,
+    resolveStateParam,
+} from '../helpers/custom-state';
+
+export const diagnostics = {
+    ...stateDiagnostics,
+};
+
+// HOOKS
+
+export const hooks = createFeature({});
+
+const delimiters = {
+    booleanStateDelimiter,
+    stateMiddleDelimiter,
+    stateWithParamDelimiter,
+};
+export {
+    parsePseudoStates,
+    transformPseudoClassToCustomState,
+    delimiters,
+    createBooleanStateClassName,
+    createStateWithParamClassName,
+    systemValidators,
+    resolveStateParam,
+};
