+++ packages/core/src/stylable-meta.ts
@@ -12,6 +12,7 @@ import {
     STScope,
     STVar,
     STCustomSelector,
+    STCustomState,
     STMixin,
     CSSClass,
     CSSType,
@@ -28,6 +29,7 @@ const features = [
     STScope,
     STVar,
     STCustomSelector,
+    STCustomState,
     STMixin,
     CSSClass,
     CSSType,
