+++ packages/core/src/types.ts
@@ -10,19 +10,6 @@ export interface ParsedValue {
     url?: string;
 }
 
-export interface StateTypeValidator {
-    name: string;
-    args: string[];
-}
-
-export type StateArguments = Array<StateTypeValidator | string>;
-
-export interface StateParsedValue {
-    type: string;
-    defaultValue?: string;
-    arguments: StateArguments;
-}
-
 export interface OptimizeConfig {
     removeComments?: boolean;
     removeStylableDirectives?: boolean;
