+++ packages/core/test/diagnostic-codes.spec.ts
@@ -11,6 +11,7 @@ import {
     STSymbol,
     STVar,
     STCustomSelector,
+    STCustomState,
 } from '@stylable/core/dist/features';
 import { generalDiagnostics } from '@stylable/core/dist/features/diagnostics';
 import { atPropertyValidationWarnings } from '@stylable/core/dist/helpers/css-custom-property';
@@ -18,7 +19,6 @@ import { parseImportMessages, ensureImportsMessages } from '@stylable/core/dist/
 import { mixinHelperDiagnostics } from '@stylable/core/dist/helpers/mixin';
 import { valueDiagnostics } from '@stylable/core/dist/helpers/value';
 import { functionDiagnostics } from '@stylable/core/dist/functions';
-import { stateDiagnostics } from '@stylable/core/dist/pseudo-states';
 import { processorDiagnostics } from '@stylable/core/dist/stylable-processor';
 import { transformerDiagnostics } from '@stylable/core/dist/stylable-transformer';
 import { utilDiagnostics, sourcePathDiagnostics } from '@stylable/core/dist/stylable-utils';
@@ -46,7 +46,7 @@ describe('diagnostics error codes', () => {
             ...mixinHelperDiagnostics,
             ...valueDiagnostics,
             ...functionDiagnostics,
-            ...stateDiagnostics,
+            ...STCustomState.diagnostics,
             ...processorDiagnostics,
             ...transformerDiagnostics,
             ...utilDiagnostics,
